AI/requirements.txt
# Image processing (install opencv-headless first to avoid GUI version)
opencv-python-headless==4.10.0.84

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.1

# Machine Learning and AI
torch==2.4.1
torchvision==0.19.1
ultralytics==8.3.27

# Image processing
Pillow==10.4.0

# Data handling
numpy==1.26.4
pydantic==2.11.7